+++ src/combinator/provided/wrap.rs
@@ -23,13 +23,13 @@ impl<T, State, Heap> Wrap<T, State, Heap> {
 }
 
 impl<
-    Kind,
+    Value,
     State,
     Heap,
-    F: for<'text> Fn(&mut Input<'text, &mut State, &mut Heap>) -> Option<Output<'text, Kind>>,
+    F: for<'text> Fn(&mut Input<'text, &mut State, &mut Heap>) -> Option<Output<'text, Value>>,
   > Parse for Wrap<F, State, Heap>
 {
-  type Kind = Kind;
+  type Value = Value;
   type State = State;
   type Heap = Heap;
 
@@ -37,7 +37,7 @@ impl<
   fn parse<'text>(
     &self,
     input: &mut Input<'text, &mut Self::State, &mut Self::Heap>,
-  ) -> Option<Output<'text, Self::Kind>> {
+  ) -> Option<Output<'text, Self::Value>> {
     (self.inner)(input)
   }
 }
@@ -45,13 +45,13 @@ impl<
 /// Wrap a closure to create a [`Combinator`].
 #[inline]
 pub const fn wrap<
-  F: for<'text> Fn(&mut Input<'text, &mut State, &mut Heap>) -> Option<Output<'text, Kind>>,
-  Kind,
+  F: for<'text> Fn(&mut Input<'text, &mut State, &mut Heap>) -> Option<Output<'text, Value>>,
+  Value,
   State,
   Heap,
 >(
   parse: F,
-) -> Combinator!(Kind, State, Heap) {
+) -> Combinator!(Value, State, Heap) {
   Combinator::new(Wrap::new(parse))
 }
 
@@ -63,12 +63,12 @@ mod tests {
   fn combinator_wrap() {
     assert_eq!(
       wrap(|input| Some(Output {
-        kind: (),
+        value: (),
         rest: &input.rest()[1..]
       }))
       .parse(&mut Input::new("123", 0, &mut (), &mut ()).unwrap()),
       Some(Output {
-        kind: (),
+        value: (),
         rest: "23"
       })
     );